Modernize the News admin widget and improve usability.
This is a full rewrite of the News widget in the administration UI, not
only porting it to the new Python and JavaScript widget classes, but
also moving all RSS feed loading client-side instead of server-side and
improving the display.

The new version no longer depends on the server-side RSS feed loading,
which didn't work if Review Board was behind a firewall or proxy server.
By loading client-side, we ensure that any user with Internet access can
see the feed.

The rendering has also improved. The News feed is now larger and is no
longer off to the side. It cleanly separates dates from headlines, and
adds more room around each row so that headlines aren't jumbled
together.

This makes use of the new reload support, allowing users to reload the
news if they choose to. There's no automatic, periodic reloading at this
point, but that may be added in a future change at the `WidgetView`
level.

The design of the widget is also generic, allowing us to potentially
reuse the JavaScript side with other RSS feeds, if we ever want to.

Testing Done:
Tested the news widget against our actual RSS feed.

Tested that all links work.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10760/