Add a CSS component for a page's main content box.
The new `.rb-c-page-content-box` component is used to display a main
content area that contains a border and padding in a desktop display but
none in a mobile display. This is a pretty standard appearance for pages
like datagrids and in the administration UI, one which has been
re-implemented in a few different places, or has been built upon the old
`.box` classes.

Currently this component isn't being used anywhere, but it's being used
in much of the new admin UI.

Testing Done:
Tested with some upcoming admin UI work, and verified the appearance was
correct in both desktop and mobile modes.

Reviewed at https://reviews.reviewboard.org/r/10787/