Add a CSS component for alerts.
This component is used to show alert boxes that indicate an error,
warning, success, or useful information to present to the user. It's
intended as a replacement for the existing "important" box type, used
in the Permission Denied and other templates, and for messages when
saving in the administration UI.

The type used for an alert will dictate the icon shown, all of which
come from Font Awesome currently.

Alerts can contain a single paragraph of text, or multiple paragraphs
(or other content). These will be kept aligned to the right of the icon.

Testing Done:
Used the alert in some upcoming Administration UI work, and put up a
few tests (shown in the screenshot on the review request).

Reviewed at https://reviews.reviewboard.org/r/10790/