Enhance validation testing and default data for the repository form.
This improves the unit tests for the repository form, supplying default
data (simplifying the data that has to be posted in each test) and
ensuring that we're always setting some data we would have already
expected to be set (for instance, setting `visible=True`).

The `visible` flag in particular was needed to perform repository
validation, which executes some code paths that are pretty important to
check. We weren't setting this before, so we weren't executing these
code paths, which impacted some in-progress work. This is now set, and
we're disabling the actual repository communication parts of the tests
through spies instead of an older `validate_repository` flag we're now
removing.

Testing Done:
Unit tests passed.

Used this along with another upcoming change and found regressions that
the test suite previously did not uncover.

Reviewed at https://reviews.reviewboard.org/r/10813/