Clear out a repository's ACLs when marking it public.
When setting ACLs on a repository (invite-only review groups and users),
saving the repository, and then making the repository
publicly-accessible to the team, the ACLs were remaining as part of the
repository's settings. This caused problems down the road in situations
such as a repository being owned by a Local Site and a user on an ACL
being removed from the site, which would trigger a hidden validation
error and prevent saving the repository.

We now clear out the ACLs whenever we make a repository public, or when
loading a public repository into the repository form. This should
prevent this sort of problem from occurring in the future, and better
meet user expectations.

Testing Done:
Tested loading and saving public repositories that had ACLs set, and
verified that the ACLs were removed.

Tested loading and saving private repositories that had ACLs set.

Unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/10810/