Relabel the repository Publicly Accessible checkbox based on settings.
The repository configuration form's "Publicly accessible" checkbox has
confused a few people in the past, with some thinking it would made a
repository available to people without access to the Local Site owning
the repository, or to those without login credentials on a server.

To address this, we're now giving the checkbox a more explicit label and
help text to better describe the effects that the setting will have.

When used with a non-public Local Site, the checkbox will say
"Accessible to all users on <sitename>", with accompanying help text
that clarifies this further.

When used on a public Local Site or without a Local Site, the checkbox
will reflect the server's site-wide login setting, saying either
"Accessible to all logged-in users" or "Accessible to everyone."

Unit tests were added to ensure that the correct text is used in each
circumstance, and to check both the configured Local Site or when
limiting the form to a Local Site.

Testing Done:
Tested each possible configuration manually.

Unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/10811/