Remove some bad repo form code for bug tracker hosting URL validation.
We have some old contributed code from 2014 that was designed to
validate the bug tracker hosting URL field in the repository form. It
depended heavily on the order in which fields were validated, which is
very fragile, and the way in which it validated the contents made some
assumptions that often do not hold true. It's also just no longer
relevant, as a large part of its purpose was ensuring that a bug tracker
form's validation was respected, and we do that now as of this upcoming
3.0.16 release.

This change removes that layer of validation, simplifying code and
preventing some of the side effects it's triggering in upcoming changes.

Testing Done:
Unit tests pass.

Verified this fixed issues with some upcoming changes.

Reviewed at https://reviews.reviewboard.org/r/10812/