Fix indirect dependency for django-braces and asana.
We have a couple dependencies which in turn have dependencies which have
versions that can conflict with Django 1.6. django-oauth-toolkit depends
on django-braces and asana depends on requests-oauthlib. In both cases,
we're adding explicit dependencies on the upstream packages and pin
versions which are known to work.

Testing Done:
Was able to set up a new virtualenv without any warnings or errors.

Reviewed at https://reviews.reviewboard.org/r/10816/