Fix a breakage in the repository form due to a bad merge.
A recent resolution of merge conflict while merging the SCMTool-specific
repository forms from 3.0 into 4.0 resulted in an old variable being
referenced. This fixes it.

It also fixes a bad deprecated jQuery call to `.toggle()`, which broke
the Show/Hide SSH Public Key link. We now use the right logic to replace
it.