Make it easier to emit deprecation warnings using our modern classes.
This adds a convenience function to
`BaseRemovedInReviewBoardVersionWarning` that wraps `warnings.warn()`,
passing in the owning class and a default stack level. This massively
simplifies the process of emitting a warning, and is one fewer import
and two fewer parameters for files.

Testing Done:
Unit tests pass.

Verified the output of warnings manually.

Reviewed at https://reviews.reviewboard.org/r/10777/