Switch Review Board to Django 1.11.
This is the culmination of years of work.

We have been on Django 1.6 since Review Board 2.0. We've been stuck on
it largely due to Django 1.7's introduction of formal database migration
support, which was incompatible with our own Django Evolution. Solving
this incompatibility was a substantial effort, not just in Django
Evolution but across Djblets, Review Board, and related models, with
many large rewrites across all modules.

Now that we've solved all of this, we're finally ready to bump our
dependency to Django 1.11. This will allow us to switch to more modern
versions of other dependencies in time, and will also enable Python 3
support (though we're still categorizing this support as experimental
for now).

It's worth pointing out that Django 3.0 is the latest version, but
starting at 2.0 they dropped Python 2 support. We still require this for
Review Board 4.0, which is why we've only gone as far as Django 1.11.

Testing Done:
Installed a new development environment and ran unit tests on both
Python 2.7 and 3.5-3.7.

Reviewed at https://reviews.reviewboard.org/r/10863/