Fix building the documentation on Django 1.11 and Python 3.
This updates the documentation building process to successfully build
docs against Django 1.11 and Python 3.

The biggest set of changes are to the `webapidocs` extension, which
wasn't at all compatible with Python 3 before. It's now been updated to
use `six` in place of incompatible value types and methods and to use
correct string types in places.

The `Makefile` now runs `evolve --execute` instead of `syncdb` when
setting up the database. The database file itself is now specified with
an absolute path, which is important for SQLite.

It also fixes up an app name in the initial fixture for the docs build
to use the modern `djblets_`-prefixed app name required in Django 1.11.

Testing Done:
Built the docs under Python 2.7 and 3.7.

Reviewed at https://reviews.reviewboard.org/r/10865/