Fix a couple Django 1.11-related issues in `rbext test`
The rbext tool had a couple of problems when trying to run an
extension's test suite, now that we've updated to the newer version of
Django. The first of these was that apps were not getting initialized
properly before attempting to discover tests; this has been fixed by
adding a call to `reviewboard.initialize()`. The second was an error
attempting to use Review Board's `collectstatic`. Extension test suites
shouldn't need this, so I've turned off the collect step from rbext.

Testing Done:
Was able to use `rbext test -m ...` for a couple different extensions.

Reviewed at https://reviews.reviewboard.org/r/10870/