Add a heading style for rb-c-alert CSS components.
This adds a `.rb-c-alert__heading` CSS class for alerts, which is used
to add a standardized header at the top of an alert or above some
content within the alert. It's entirely optional, but ensures that we
have some structure for this common use case.

Testing Done:
Made use of this in an upcoming change, both for a leading header
and for a subsequent header within content.

Reviewed at https://reviews.reviewboard.org/r/10875/