Make improvements to RB.DialogView
This review request changes makes improvements to `RB.DialogView.` It
introduces a property on `RB.DialogView` which is a map of buttons in
the dialog from button `id`s to button elements. The `id` field of a
`RB.DialogView` button is now required so that buttons can be referenced
by their ID in the map. In addition, `disabled` and `class` are added as
button keys so that buttons can be created with initial state set to
disabled and with classes set, respectively.

Testing Done:
Unit tests:
- new dialog view button properties (id, class, disabled)

Reviewed at https://reviews.reviewboard.org/r/10820/