Update the Change Form template for improved form alert support.
This rearranges the base template for the Change Form to prepare for
some updates to some of the settings and model pages.

The `<form>` is now the main container for the page's contents, and
contains the alerts (such as for form errors) and the content box. This
inverses the relationship these elements had before, in order to allow
alerts to more easily interact with the form (providing additional form
submission buttons, for instance).

The form alerts now have their own block called `form-alerts`,
positioned above the form content box. Templates that extend this
template can more easily add their own alerts in a standard place.

Testing Done:
Tested along with some upcoming work on the repository configuration page.

Reviewed at https://reviews.reviewboard.org/r/10876/