Update the presentation of the repository configuration page.
This updates the styling and templates for the page to leverage the new
CSS component classes for the form and the confirmation prompts,
finishing up the presentation of this page.

The form and fieldsets now match the new admin UI style, and the
fieldsets are injected via a specialization of the main fieldset
template.

The confirmation prompts and SSH key info now live in `rb-c-alert`
components, with new admin-specific CSS component classes for the SSH
key and certificate information.

Testing Done:
Tested all the various confirmation prompts for the form, including
buttons when shown.

Tested creating plain and service-backed repositories.

Tested with and without SSH keys on the system.

Reviewed at https://reviews.reviewboard.org/r/10885/